% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate10.R
\name{translate10}
\alias{translate10}
\title{Translate10}
\usage{
translate10(icd, q)
}
\arguments{
\item{icd}{The ICD10 Code}

\item{q}{the thing to look for}
}
\value{
Disorder,ICD9
}
\description{
Translate ICD-10 to ICD-9 and name of DSM-5 disorder
}
\examples{
    translate10(icd,"E66.9")
}
