% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate9.R
\name{translate9}
\alias{translate9}
\title{Translate9}
\usage{
translate9(icd, q)
}
\arguments{
\item{icd}{The ICD-9 Code}

\item{q}{the thing to look for}
}
\value{
Disorder,ICD10
}
\description{
Translate ICD-9 code to ICD-10 code and name of DSM-5 disorder
}
\examples{
    translate9(icd,"278")
}
